requirements.txt
hopsworks
pandas
numpy
scikit-learn
python-dotenv
requests
joblib
streamlit
matplotlib
